refactor: restructure test organization and consolidate duplicate tests
This commit addresses Issue #41 by implementing a comprehensive test structure refactoring:

## Key Changes:
- **Removed unused testing/ directory** that contained outdated scripts and HTML files
- **Consolidated duplicate test files** from 16 files down to 3 comprehensive test files
- **Standardized test organization** under tests/unit/services/ structure
- **Fixed all test failures** and ensured 100% test pass rate (109 passing tests)

## Test Consolidations:
- **Minecraft Server**: 4 duplicate files → 1 comprehensive test (63% coverage)
- **Backup Service**: 2 duplicate files → 1 comprehensive test (40% coverage)
- **Group Service**: 3 duplicate files → 1 comprehensive test (39% coverage)

## Files Removed:
- testing/ directory (4 files)
- 13 duplicate/obsolete test files

## Files Added/Updated:
- tests/unit/services/test_minecraft_server.py (consolidated)
- tests/unit/services/test_backup_service.py (consolidated)
- tests/unit/services/test_group_service.py (consolidated)

The refactored test structure improves maintainability, reduces duplication, and provides better organization while maintaining comprehensive test coverage.
