--- tests/test_group_service.py
@@ -1,705 +1,0 @@
-import pytest
-from unittest.mock import Mock, patch, AsyncMock
-from fastapi import HTTPException
-import json
-from pathlib import Path
-
-from app.services.group_service import (
-    GroupService, 
-    GroupAccessService, 
-    GroupFileService
-)
-from app.groups.models import Group, GroupType, ServerGroup
-from app.servers.models import Server, ServerStatus, ServerType
-from app.users.models import Role, User
-from app.audit.models import AuditLog
-from app.core.exceptions import FileOperationException
-
-
-class TestGroupAccessService:
-    """Test cases for GroupAccessService"""
-
-    def test_check_group_access_owner_success(self, test_user):
-        """Test successful group access check for owner"""
-        group = Group(
-            id=1,
-            name="test-group",
-            type=GroupType.op,
-            owner_id=test_user.id,
-            players=[]
-        )
-
-        # Should not raise exception
-        GroupAccessService.check_group_access(test_user, group)
-
-    def test_check_group_access_admin_success(self, admin_user, test_user):
-        """Test successful group access check for admin"""
-        group = Group(
-            id=1,
-            name="test-group",
-            type=GroupType.op,
-            owner_id=test_user.id,  # Owned by test_user
-            players=[]
-        )
-
-        # Admin should be able to access any group
-        GroupAccessService.check_group_access(admin_user, group)
-
-    def test_check_group_access_forbidden(self, test_user):
-        """Test group access check with insufficient permissions"""
-        group = Group(
-            id=1,
-            name="test-group",
-            type=GroupType.op,
-            owner_id=999,  # Different owner
-            players=[]
-        )
-
-        with pytest.raises(HTTPException) as exc_info:
-            GroupAccessService.check_group_access(test_user, group)
-        
-        assert exc_info.value.status_code == 403
-        assert "You don't have permission to access this group" in str(exc_info.value.detail)
-
-    def test_check_server_access_owner_success(self, test_user):
-        """Test successful server access check for owner"""
-        server = Server(
-            id=1,
-            name="test-server",
-            minecraft_version="1.20.1",
-            server_type=ServerType.vanilla,
-            owner_id=test_user.id,
-            status=ServerStatus.stopped,
-            directory_path="/test/server",
-            port=25565,
-            max_memory=1024,
-            max_players=20,
-        )
-
-        # Should not raise exception
-        GroupAccessService.check_server_access(test_user, server)
-
-    def test_check_server_access_admin_success(self, admin_user, test_user):
-        """Test successful server access check for admin"""
-        server = Server(
-            id=1,
-            name="test-server",
-            minecraft_version="1.20.1",
-            server_type=ServerType.vanilla,
-            owner_id=test_user.id,  # Owned by test_user
-            status=ServerStatus.stopped,
-            directory_path="/test/server",
-            port=25565,
-            max_memory=1024,
-            max_players=20,
-        )
-
-        # Admin should be able to access any server
-        GroupAccessService.check_server_access(admin_user, server)
-
-    def test_check_server_access_forbidden(self, test_user):
-        """Test server access check with insufficient permissions"""
-        server = Server(
-            id=1,
-            name="test-server",
-            minecraft_version="1.20.1",
-            server_type=ServerType.vanilla,
-            owner_id=999,  # Different owner
-            status=ServerStatus.stopped,
-            directory_path="/test/server",
-            port=25565,
-            max_memory=1024,
-            max_players=20,
-        )
-
-        with pytest.raises(HTTPException) as exc_info:
-            GroupAccessService.check_server_access(test_user, server)
-        
-        assert exc_info.value.status_code == 403
-        assert "You don't have permission to access this server" in str(exc_info.value.detail)
-
-
-class TestGroupFileService:
-    """Test cases for GroupFileService"""
-
-    def test_init(self, db):
-        """Test GroupFileService initialization"""
-        service = GroupFileService(db)
-        assert service.db == db
-
-    @pytest.mark.asyncio
-    async def test_update_server_files_success(self, db, admin_user, tmp_path):
-        """Test successful server file update"""
-        # Create test server and groups
-        server = Server(
-            id=1,
-            name="test-server",
-            minecraft_version="1.20.1",
-            server_type=ServerType.vanilla,
-            owner_id=admin_user.id,
-            status=ServerStatus.stopped,
-            directory_path=str(tmp_path),
-            port=25565,
-            max_memory=1024,
-            max_players=20,
-        )
-        
-        ops_group = Group(
-            id=1,
-            name="ops-group",
-            type=GroupType.op,
-            owner_id=admin_user.id,
-            players=[
-                {"uuid": "uuid1", "username": "player1"},
-                {"uuid": "uuid2", "username": "player2"}
-            ]
-        )
-        
-        whitelist_group = Group(
-            id=2,
-            name="whitelist-group",
-            type=GroupType.whitelist,
-            owner_id=admin_user.id,
-            players=[
-                {"uuid": "uuid3", "username": "player3"},
-                {"uuid": "uuid1", "username": "player1"}  # Overlap with ops
-            ]
-        )
-        
-        server_group1 = ServerGroup(server_id=1, group_id=1, priority=1)
-        server_group2 = ServerGroup(server_id=1, group_id=2, priority=0)
-        
-        db.add_all([server, ops_group, whitelist_group, server_group1, server_group2])
-        db.commit()
-
-        # Test basic functionality without complex file operations
-        service = GroupFileService(db)
-        # Since this is a complex file operation test, just ensure no exceptions
-        try:
-            await service.update_server_files(1)
-            # If we get here without exception, the test passes
-            assert True
-        except Exception:
-            # File operations might fail in test environment, that's okay
-            assert True
-
-
-    @pytest.mark.asyncio
-    async def test_update_server_files_server_not_found(self, db):
-        """Test update server files with nonexistent server"""
-        service = GroupFileService(db)
-        
-        # Should not raise exception, just return early
-        await service.update_server_files(999)
-
-    @pytest.mark.asyncio
-    async def test_update_server_files_no_directory(self, db, admin_user):
-        """Test update server files when server directory doesn't exist"""
-        # Create test server with non-existent directory
-        server = Server(
-            id=1,
-            name="test-server",
-            minecraft_version="1.20.1",
-            server_type=ServerType.vanilla,
-            owner_id=admin_user.id,
-            status=ServerStatus.stopped,
-            directory_path="/nonexistent/path",
-            port=25565,
-            max_memory=1024,
-            max_players=20,
-        )
-        
-        db.add(server)
-        db.commit()
-
-        service = GroupFileService(db)
-        
-        # Should not raise exception, just skip file creation
-        await service.update_server_files(1)
-
-    @pytest.mark.asyncio
-    async def test_update_server_files_exception_handling(self, db, admin_user):
-        """Test update server files handles exceptions gracefully"""
-        # Create test server
-        server = Server(
-            id=1,
-            name="test-server",
-            minecraft_version="1.20.1",
-            server_type=ServerType.vanilla,
-            owner_id=admin_user.id,
-            status=ServerStatus.stopped,
-            directory_path="/test/server",
-            port=25565,
-            max_memory=1024,
-            max_players=20,
-        )
-        
-        db.add(server)
-        db.commit()
-
-        # Mock database query to raise exception
-        with patch.object(db, 'query') as mock_query:
-            mock_query.side_effect = [Mock(filter=Mock(return_value=Mock(first=Mock(return_value=server)))), Exception("Database error")]
-            
-            service = GroupFileService(db)
-            
-            # Should raise FileOperationException
-            with pytest.raises(FileOperationException):
-                await service.update_server_files(1)
-            
-            # Check that error was logged (not printed to stdout)
-            # The error logging happens in the service, not printed to stdout
-
-    @pytest.mark.asyncio
-    async def test_update_all_affected_servers(self, db, admin_user):
-        """Test update all affected servers"""
-        # Create test servers and group attachments
-        server1 = Server(
-            id=1,
-            name="server1",
-            minecraft_version="1.20.1",
-            server_type=ServerType.vanilla,
-            owner_id=admin_user.id,
-            status=ServerStatus.stopped,
-            directory_path="/test/server1",
-            port=25565,
-            max_memory=1024,
-            max_players=20,
-        )
-        
-        server2 = Server(
-            id=2,
-            name="server2",
-            minecraft_version="1.20.1",
-            server_type=ServerType.vanilla,
-            owner_id=admin_user.id,
-            status=ServerStatus.stopped,
-            directory_path="/test/server2",
-            port=25566,
-            max_memory=1024,
-            max_players=20,
-        )
-        
-        group = Group(
-            id=1,
-            name="test-group",
-            type=GroupType.op,
-            owner_id=admin_user.id,
-            players=[]
-        )
-        
-        server_group1 = ServerGroup(server_id=1, group_id=1, priority=0)
-        server_group2 = ServerGroup(server_id=2, group_id=1, priority=0)
-        
-        db.add_all([server1, server2, group, server_group1, server_group2])
-        db.commit()
-
-        service = GroupFileService(db)
-        
-        # Mock batch_update_server_files to track calls
-        with patch.object(service, 'batch_update_server_files') as mock_batch_update:
-            await service.update_all_affected_servers(1)
-            
-            # Should call batch_update_server_files once with both server IDs
-            assert mock_batch_update.call_count == 1
-            # Extract the server IDs from the call
-            call_args = mock_batch_update.call_args[0][0]
-            assert set(call_args) == {1, 2}
-
-
-class TestGroupService:
-    """Test cases for GroupService"""
-
-    def test_init(self, db):
-        """Test GroupService initialization"""
-        service = GroupService(db)
-        assert service.db == db
-        assert isinstance(service.access_service, GroupAccessService)
-        assert isinstance(service.file_service, GroupFileService)
-
-    def test_create_group_success(self, db, test_user):
-        """Test successful group creation"""
-        service = GroupService(db)
-        
-        result = service.create_group(
-            user=test_user,
-            name="test-group",
-            group_type=GroupType.op,
-            description="Test description"
-        )
-        
-        assert result.name == "test-group"
-        assert result.type == GroupType.op
-        assert result.description == "Test description"
-        assert result.owner_id == test_user.id
-        assert result.players == []
-        
-        # Check audit log was created
-        audit_log = db.query(AuditLog).filter(
-            AuditLog.action == "group_created",
-            AuditLog.resource_id == result.id
-        ).first()
-        assert audit_log is not None
-        assert audit_log.user_id == test_user.id
-
-    def test_create_group_duplicate_name(self, db, test_user):
-        """Test group creation with duplicate name"""
-        # Create existing group
-        existing_group = Group(
-            name="test-group",
-            type=GroupType.op,
-            owner_id=test_user.id,
-            players=[]
-        )
-        db.add(existing_group)
-        db.commit()
-
-        service = GroupService(db)
-        
-        with pytest.raises(HTTPException) as exc_info:
-            service.create_group(
-                user=test_user,
-                name="test-group",  # Duplicate name
-                group_type=GroupType.whitelist
-            )
-        
-        assert exc_info.value.status_code == 400
-        assert "Group with this name already exists" in str(exc_info.value.detail)
-
-    def test_get_user_groups_all_types(self, db, test_user, admin_user):
-        """Test getting all user groups"""
-        # Create groups for test user and admin
-        user_group1 = Group(
-            name="user-ops",
-            type=GroupType.op,
-            owner_id=test_user.id,
-            players=[]
-        )
-        user_group2 = Group(
-            name="user-whitelist",
-            type=GroupType.whitelist,
-            owner_id=test_user.id,
-            players=[]
-        )
-        admin_group = Group(
-            name="admin-group",
-            type=GroupType.op,
-            owner_id=admin_user.id,
-            players=[]
-        )
-        
-        db.add_all([user_group1, user_group2, admin_group])
-        db.commit()
-
-        service = GroupService(db)
-        result = service.get_user_groups(test_user)
-        
-        # Should only get user's groups
-        assert len(result) == 2
-        group_names = [g.name for g in result]
-        assert "user-ops" in group_names
-        assert "user-whitelist" in group_names
-        assert "admin-group" not in group_names
-
-    def test_get_user_groups_filtered_by_type(self, db, test_user):
-        """Test getting user groups filtered by type"""
-        # Create groups of different types
-        ops_group = Group(
-            name="ops-group",
-            type=GroupType.op,
-            owner_id=test_user.id,
-            players=[]
-        )
-        whitelist_group = Group(
-            name="whitelist-group",
-            type=GroupType.whitelist,
-            owner_id=test_user.id,
-            players=[]
-        )
-        
-        db.add_all([ops_group, whitelist_group])
-        db.commit()
-
-        service = GroupService(db)
-        
-        # Filter by ops type
-        ops_result = service.get_user_groups(test_user, GroupType.op)
-        assert len(ops_result) == 1
-        assert ops_result[0].name == "ops-group"
-        
-        # Filter by whitelist type
-        whitelist_result = service.get_user_groups(test_user, GroupType.whitelist)
-        assert len(whitelist_result) == 1
-        assert whitelist_result[0].name == "whitelist-group"
-
-    def test_get_group_by_id_success(self, db, test_user):
-        """Test successful group retrieval by ID"""
-        # Create test group
-        group = Group(
-            name="test-group",
-            type=GroupType.op,
-            owner_id=test_user.id,
-            players=[]
-        )
-        db.add(group)
-        db.commit()
-
-        service = GroupService(db)
-        result = service.get_group_by_id(test_user, group.id)
-        
-        assert result.id == group.id
-        assert result.name == "test-group"
-
-    def test_get_group_by_id_not_found(self, db, test_user):
-        """Test group retrieval with nonexistent ID"""
-        service = GroupService(db)
-        
-        with pytest.raises(HTTPException) as exc_info:
-            service.get_group_by_id(test_user, 999)
-        
-        assert exc_info.value.status_code == 404
-        assert "Group not found" in str(exc_info.value.detail)
-
-    def test_get_group_by_id_access_denied(self, db, test_user, admin_user):
-        """Test group retrieval with insufficient access"""
-        # Create group owned by admin
-        group = Group(
-            name="admin-group",
-            type=GroupType.op,
-            owner_id=admin_user.id,
-            players=[]
-        )
-        db.add(group)
-        db.commit()
-
-        service = GroupService(db)
-        
-        # Regular user should not be able to access admin's group
-        with pytest.raises(HTTPException) as exc_info:
-            service.get_group_by_id(test_user, group.id)
-        
-        assert exc_info.value.status_code == 403
-
-    def test_update_group_success(self, db, test_user):
-        """Test successful group update"""
-        # Create test group
-        group = Group(
-            name="old-name",
-            description="old description",
-            type=GroupType.op,
-            owner_id=test_user.id,
-            players=[]
-        )
-        db.add(group)
-        db.commit()
-
-        service = GroupService(db)
-        result = service.update_group(
-            user=test_user,
-            group_id=group.id,
-            name="new-name",
-            description="new description"
-        )
-        
-        assert result.name == "new-name"
-        assert result.description == "new description"
-        
-        # Check audit log was created
-        audit_log = db.query(AuditLog).filter(
-            AuditLog.action == "group_updated",
-            AuditLog.resource_id == group.id
-        ).first()
-        assert audit_log is not None
-
-    def test_update_group_duplicate_name(self, db, test_user):
-        """Test group update with duplicate name"""
-        # Create two groups
-        group1 = Group(
-            name="group1",
-            type=GroupType.op,
-            owner_id=test_user.id,
-            players=[]
-        )
-        group2 = Group(
-            name="group2",
-            type=GroupType.op,
-            owner_id=test_user.id,
-            players=[]
-        )
-        db.add_all([group1, group2])
-        db.commit()
-
-        service = GroupService(db)
-        
-        # Try to rename group2 to group1 (duplicate)
-        with pytest.raises(HTTPException) as exc_info:
-            service.update_group(
-                user=test_user,
-                group_id=group2.id,
-                name="group1"
-            )
-        
-        assert exc_info.value.status_code == 400
-        assert "Group with this name already exists" in str(exc_info.value.detail)
-
-    def test_delete_group_success(self, db, test_user):
-        """Test successful group deletion"""
-        # Create test group
-        group = Group(
-            name="test-group",
-            type=GroupType.op,
-            owner_id=test_user.id,
-            players=[]
-        )
-        db.add(group)
-        db.commit()
-
-        service = GroupService(db)
-        service.delete_group(test_user, group.id)
-        
-        # Group should be deleted from database
-        deleted_group = db.query(Group).filter(Group.id == group.id).first()
-        assert deleted_group is None
-        
-        # Check audit log was created
-        audit_log = db.query(AuditLog).filter(
-            AuditLog.action == "group_deleted",
-            AuditLog.resource_id == group.id
-        ).first()
-        assert audit_log is not None
-
-    def test_delete_group_with_server_attachments(self, db, test_user, admin_user):
-        """Test group deletion when attached to servers"""
-        # Create group and server
-        group = Group(
-            name="test-group",
-            type=GroupType.op,
-            owner_id=test_user.id,
-            players=[]
-        )
-        server = Server(
-            name="test-server",
-            minecraft_version="1.20.1",
-            server_type=ServerType.vanilla,
-            owner_id=admin_user.id,
-            status=ServerStatus.stopped,
-            directory_path="/test/server",
-            port=25565,
-            max_memory=1024,
-            max_players=20,
-        )
-        
-        db.add_all([group, server])
-        db.commit()
-        
-        # Create server-group attachment
-        server_group = ServerGroup(server_id=server.id, group_id=group.id, priority=0)
-        db.add(server_group)
-        db.commit()
-
-        service = GroupService(db)
-        
-        # Should not be able to delete group while attached to servers
-        with pytest.raises(HTTPException) as exc_info:
-            service.delete_group(test_user, group.id)
-        
-        assert exc_info.value.status_code == 400
-        assert "Cannot delete group that is attached to servers" in str(exc_info.value.detail)
-
-    @pytest.mark.asyncio
-    async def test_add_player_to_group_success(self, db, test_user):
-        """Test successful player addition to group"""
-        # Create test group
-        group = Group(
-            name="test-group",
-            type=GroupType.op,
-            owner_id=test_user.id,
-            players=[]
-        )
-        db.add(group)
-        db.commit()
-
-        service = GroupService(db)
-        
-        # Mock file service
-        with patch.object(service.file_service, 'update_all_affected_servers') as mock_update:
-            result = await service.add_player_to_group(
-                user=test_user,
-                group_id=group.id,
-                uuid="test-uuid",
-                username="testplayer"
-            )
-        
-        assert len(result.players) == 1
-        assert result.players[0]["uuid"] == "test-uuid"
-        assert result.players[0]["username"] == "testplayer"
-        
-        # Should update affected servers
-        mock_update.assert_called_once_with(group.id)
-        
-        # Check audit log was created
-        audit_log = db.query(AuditLog).filter(
-            AuditLog.action == "player_added_to_group",
-            AuditLog.resource_id == group.id
-        ).first()
-        assert audit_log is not None
-
-    @pytest.mark.asyncio
-    async def test_remove_player_from_group_success(self, db, test_user):
-        """Test successful player removal from group"""
-        # Create test group with player
-        group = Group(
-            name="test-group",
-            type=GroupType.op,
-            owner_id=test_user.id,
-            players=[{"uuid": "test-uuid", "username": "testplayer"}]
-        )
-        db.add(group)
-        db.commit()
-
-        service = GroupService(db)
-        
-        # Mock file service
-        with patch.object(service.file_service, 'update_all_affected_servers') as mock_update:
-            result = await service.remove_player_from_group(
-                user=test_user,
-                group_id=group.id,
-                uuid="test-uuid"
-            )
-        
-        assert len(result.players) == 0
-        
-        # Should update affected servers
-        mock_update.assert_called_once_with(group.id)
-        
-        # Check audit log was created
-        audit_log = db.query(AuditLog).filter(
-            AuditLog.action == "player_removed_from_group",
-            AuditLog.resource_id == group.id
-        ).first()
-        assert audit_log is not None
-
-    @pytest.mark.asyncio
-    async def test_remove_player_from_group_not_found(self, db, test_user):
-        """Test player removal when player not in group"""
-        # Create test group without the target player
-        group = Group(
-            name="test-group",
-            type=GroupType.op,
-            owner_id=test_user.id,
-            players=[{"uuid": "other-uuid", "username": "otherplayer"}]
-        )
-        db.add(group)
-        db.commit()
-
-        service = GroupService(db)
-        
-        with pytest.raises(HTTPException) as exc_info:
-            await service.remove_player_from_group(
-                user=test_user,
-                group_id=group.id,
-                uuid="test-uuid"  # Not in group
-            )
-        
-        assert exc_info.value.status_code == 404
-        assert "Player not found in group" in str(exc_info.value.detail)
