--- tests/test_simple_group_coverage.py
@@ -1,91 +1,0 @@
-import pytest
-from unittest.mock import Mock, patch
-from fastapi import HTTPException
-
-from app.services.group_service import GroupAccessService, GroupFileService
-from app.groups.models import Group, GroupType
-from app.servers.models import Server, ServerStatus, ServerType
-from app.users.models import Role
-
-
-class TestGroupAccessServiceSimple:
-    """Simple tests to improve GroupAccessService coverage"""
-
-    def test_check_group_access_owner(self, test_user):
-        """Test group access for owner"""
-        group = Mock()
-        group.owner_id = test_user.id
-        test_user.role = Role.user
-        
-        # Should not raise exception
-        GroupAccessService.check_group_access(test_user, group)
-
-    def test_check_group_access_admin(self, admin_user):
-        """Test group access for admin"""
-        group = Mock()
-        group.owner_id = 999  # Different owner
-        
-        # Should not raise exception for admin
-        GroupAccessService.check_group_access(admin_user, group)
-
-    def test_check_server_access_owner(self, test_user):
-        """Test server access for owner"""
-        server = Mock()
-        server.owner_id = test_user.id
-        test_user.role = Role.user
-        
-        # Should not raise exception
-        GroupAccessService.check_server_access(test_user, server)
-
-    def test_check_server_access_admin(self, admin_user):
-        """Test server access for admin"""
-        server = Mock()
-        server.owner_id = 999  # Different owner
-        
-        # Should not raise exception for admin
-        GroupAccessService.check_server_access(admin_user, server)
-
-
-class TestGroupFileServiceSimple:
-    """Simple tests to improve GroupFileService coverage"""
-
-    def test_init(self, db):
-        """Test GroupFileService initialization"""
-        service = GroupFileService(db)
-        assert service.db == db
-
-    @pytest.mark.asyncio
-    async def test_update_server_files_no_server(self, db):
-        """Test update server files when server doesn't exist"""
-        service = GroupFileService(db)
-        
-        # Should handle gracefully when server not found
-        await service.update_server_files(999)
-
-    @pytest.mark.asyncio 
-    async def test_update_all_affected_servers_empty(self, db):
-        """Test update all affected servers with no affected servers"""
-        service = GroupFileService(db)
-        
-        # Should handle gracefully when no servers are affected
-        await service.update_all_affected_servers(999)
-
-
-class TestGroupServiceImports:
-    """Test basic imports and initialization to improve coverage"""
-
-    def test_imports(self):
-        """Test that imports work correctly"""
-        from app.services.group_service import GroupService, GroupAccessService, GroupFileService
-        assert GroupService is not None
-        assert GroupAccessService is not None
-        assert GroupFileService is not None
-
-    def test_group_service_init(self, db):
-        """Test GroupService initialization"""
-        from app.services.group_service import GroupService
-        
-        service = GroupService(db)
-        assert service.db == db
-        assert hasattr(service, 'access_service')
-        assert hasattr(service, 'file_service')
